Add main README, license, and markdown script
Defines documentation partials to generate the main repository readme.
This allows for an easier editing experience and reusable documentation
if needed.

Current license is a 3-Clause BSD until a more appropriate license can
be decided upon. If a depedencies license is incompatible with the
chosen license, the dependencies lices supersedes the chosen license for
affected files/areas.

The markdown.js script handles automatically generating a table of
contents for a markdown file, as well as inlining partials to modularize
documentation. This is inspired by IA Writer from Information Architects