Fix replace post with get for sources
Following REST practices with the name, we want to read items, not
create, POST does not make sense in this case. This gives the
opportunity to expand on some of the middleware bringing query params
and more normalization operations.

Reordering the logger request middleware up allows us to see the request
before it has a chance to fail on the middleware. Minor, but helpful.